pyproject.toml
[project]
name = "autoapply"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.59.0",
    "nltk>=3.9.2",
    "numpy>=2.4.1",
    "playwright>=1.57.0",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
